Method Transfer
From MainActivity to ConnectionDetector
